+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * :sparkles: Network add getLocalIpV4Address
 * :arrow_up: minSdkVersion rose to 16
 * :sparkles: Add StorageManagerCompat and StorageVolumeCompat
+* :hammer: StatFs getCompatAvailableBytes, getCompatFreeBytes, getCompatTotalBytes rename to getAvailableBytesCompat, getFreeBytesCompat, getTotalBytesCompat
 
 ## v0.2.2
 * Upgrade Javax 0.6.4